Improved (MainWindow::DeleteAllBooksFromGroup()): Сделан запрос на согласие удаления всех книг выделенной группы из структур текущей библиотеки.

+++ freeLib/src/mainwindow.cpp
@@ -3340,35 +3340,40 @@ void MainWindow::AddBookToGroupAction()
 */
 void MainWindow::DeleteAllBooksFromGroup()
 {
-    // Формирование списка книг для выделенной Группы
-    QList<uint> listBooks = MakeListBooksFromSelectedGroup(g_idCurrentLib);
-    
-    // удаление в базе книг из сформированного списка
-    QSqlQuery query(QSqlDatabase::database("libdb"));
-    foreach(uint book_id, listBooks) {
-        query.prepare("DELETE FROM book_group WHERE id_lib = :id_lib AND group_id = :group_id AND book_id = :book_id;");
-        query.bindValue(":book_id", book_id);
-        query.bindValue(":group_id", idCurrentGroup_);
-        query.bindValue(":id_lib", g_idCurrentLib);
-        if (!query.exec())
-            qDebug() << query.lastError().text();
-    }
+    if (QMessageBox::question(
+        this, tr("Clear selected group"),
+        tr("Are you sure you want to delete all books of the selected group?"),
+        QMessageBox::Yes | QMessageBox::No, QMessageBox::No) == QMessageBox::Yes) {
+        // Формирование списка книг для выделенной Группы
+        QList<uint> listBooks = MakeListBooksFromSelectedGroup(g_idCurrentLib);
 
-    // удаление из структуры связи этой книги с выделенной группой
-    QHash<uint, SBook>::iterator BookIterator = mLibs[g_idCurrentLib].mBooks.begin();
-    while (BookIterator != mLibs[g_idCurrentLib].mBooks.end()) {
-        if (idCurrentLanguage_ == -1 || idCurrentLanguage_ == BookIterator->idLanguage) {
-            QMutableListIterator<uint> GroupIterator(BookIterator->listIdGroups);
-            while (GroupIterator.hasNext()) {
-                if (GroupIterator.next() == idCurrentGroup_) {
-                    GroupIterator.remove();
-                    break;
+        // удаление в базе книг из сформированного списка
+        QSqlQuery query(QSqlDatabase::database("libdb"));
+        foreach(uint book_id, listBooks) {
+            query.prepare("DELETE FROM book_group WHERE id_lib = :id_lib AND group_id = :group_id AND book_id = :book_id;");
+            query.bindValue(":book_id", book_id);
+            query.bindValue(":group_id", idCurrentGroup_);
+            query.bindValue(":id_lib", g_idCurrentLib);
+            if (!query.exec())
+                qDebug() << query.lastError().text();
+        }
+
+        // удаление из структуры связи этой книги с выделенной группой
+        QHash<uint, SBook>::iterator BookIterator = mLibs[g_idCurrentLib].mBooks.begin();
+        while (BookIterator != mLibs[g_idCurrentLib].mBooks.end()) {
+            if (idCurrentLanguage_ == -1 || idCurrentLanguage_ == BookIterator->idLanguage) {
+                QMutableListIterator<uint> GroupIterator(BookIterator->listIdGroups);
+                while (GroupIterator.hasNext()) {
+                    if (GroupIterator.next() == idCurrentGroup_) {
+                        GroupIterator.remove();
+                        break;
+                    }
                 }
             }
+            ++BookIterator;
         }
-        ++BookIterator;
+        ui->Books->clear();
     }
-    ui->Books->clear();
 }
 
 /*
